feat: implement historical data access
This adds historical data capabilities. The game will save the history of played games in a PostgreSQL database as it is played, and on each win tell the player how many games they have played in the past hour, and what the percentage of won games was. We are running a morally sound establishment here!